+++ core/tests/functional/gallery_lifecycle.php
@@ -154,6 +154,7 @@ class gallery_lifecycle extends \phpbb_functional_test_case
 		$db = $this->get_db();
 		$sql_ary = [
 			'parent_id'        => 0,
+			'album_parents'    => '',
 			'left_id'          => 1,
 			'right_id'         => 2,
 			'album_type'       => \phpbbgallery\core\block::TYPE_UPLOAD,