+++ core/tests/functional/image_move_links_workflow.php
@@ -113,6 +113,7 @@ class image_move_links_workflow extends \phpbb_functional_test_case
 		$db = $this->get_db();
 		$db->sql_query('INSERT INTO phpbb_gallery_albums ' . $db->sql_build_array('INSERT', [
 			'parent_id' => 0,
+			'album_parents' => '',
 			'left_id' => $left_id,
 			'right_id' => $left_id + 1,
 			'album_type' => \phpbbgallery\core\block::TYPE_UPLOAD,