+++ core/tests/functional/notification_privacy_workflow.php
@@ -126,6 +126,7 @@ class notification_privacy_workflow extends \phpbb_functional_test_case
 		$db = $this->get_db();
 		$db->sql_query('INSERT INTO phpbb_gallery_albums ' . $db->sql_build_array('INSERT', [
 			'parent_id' => 0,
+			'album_parents' => '',
 			'left_id' => 1,
 			'right_id' => 2,
 			'album_type' => \phpbbgallery\core\block::TYPE_UPLOAD,